Added parser and LS benchmarks
The benchmarks are designed so that they generate an input string, for a given minimum length.

The use a number of fragments picked from the code bases of these languages.

The Base benchmarks measure parser speed.
In the case of Simple, that’s the generated parser, without generating any details, just parsing.
For Som and Newspeak, it’s the baseline AST parsers, so, they generate already the AST, but do not collect any LS information.

The Ls benchmarks use the LS infrastructure, so, here we see the overhead of collecting the information.

The Retrieve benchmarks on top of that retrieve symbol and token information, which requires a little bit of extra work, but it doesn’t actually serialize the data. So, not a true end-to-end benchmark.

Signed-off-by: Stefan Marr <git@stefan-marr.de>